slipshot_backend/requirements.txt
# Django & REST Framework
Django>=5.2.10
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.9.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# Production Server
gunicorn>=23.0.0
whitenoise>=6.7.0

# Environment Variables
python-dotenv>=1.2.0

# Image Processing & OCR
pillow>=12.0.0
easyocr>=1.7.0
pytesseract>=0.3.10
opencv-python-headless>=4.10.0
pyzbar>=0.1.9

# ML/Deep Learning (for EasyOCR)
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=2.0.0
scipy>=1.14.0
scikit-image>=0.24.0

# Other
PyYAML>=6.0.0
